requirements.txt
dateutils==0.6.6
django-extensions==1.7.5
django-filter==1.0.4
django-froala-editor==2.4.2
django-measurement==2.4.0
django-multiselectfield==0.1.8
Django<2.0
graphene-django>=2.0.0
graphene==2.0.1
isoweek==1.3.3
jsonfield==2.0.2
mailchimp3==2.0.11
Markdown==2.6.7
measurement==1.8.0
nameparser==0.5.1
phonenumbers==8.8.8
Pillow==4.0.0
probableparsing==0.0.1
py3dns==3.1.0
pyDes==2.0.1
PyJWT==1.5.3
python-crfsuite==0.9.5
python-dateutil==2.6.1
pytz==2017.2
qrcode==5.3
raven==6.5.0
requests==2.13.0
simplejson==3.10.0
stripe==1.46.0
usaddress==0.5.10
validate-email==1.3
